Update main - Copy.rs
On second thought never mind. We simply filtered out immobile hosts that are not infected and the computation got significantly less expensive.

In other words, concept of time or contact wise colonization will not at all be considered for faeces or eggs - i.e. infected faces and infected eggs are not going to be modeled to be "re-infected" like hosts will be